[3D Image] Add pan & zoom support
