+++ packages/studio-base/src/i18n/en/threeDee.ts
@@ -75,4 +75,5 @@ export const threeDee = {
 
   // Image annotations
   imageAnnotations: "Image annotations",
+  resetView: "Reset view",
 };
