+++ packages/studio-base/src/panels/ThreeDeeRender/IRenderer.ts
@@ -50,6 +50,7 @@ export type RendererEvents = {
   configChange: (renderer: IRenderer) => void;
   schemaHandlersChanged: (renderer: IRenderer) => void;
   topicHandlersChanged: (renderer: IRenderer) => void;
+  resetViewChanged: (renderer: IRenderer) => void;
 };
 
 export type FollowMode = "follow-pose" | "follow-position" | "follow-none";
@@ -289,6 +290,11 @@ export interface IRenderer extends EventEmitter<RendererEvents> {
   setCameraState(cameraState: CameraState): void;
 
   getCameraState(): CameraState | undefined;
+
+  /** Whether the view has been modified and a reset button should be shown (image mode only). */
+  canResetView(): boolean;
+  /** Reset any manual view modifications (image mode only). */
+  resetView(): void;
 
   setSelectedRenderable(selection: PickedRenderable | undefined): void;
