+++ packages/studio-base/src/panels/ThreeDeeRender/stories/ImageMode/ImageMode.stories.tsx
@@ -3,7 +3,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import { StoryObj } from "@storybook/react";
-import { screen, userEvent } from "@storybook/testing-library";
+import { fireEvent, screen, userEvent } from "@storybook/testing-library";
 
 import { CompressedImage, RawImage } from "@foxglove/schemas";
 import { MessageEvent } from "@foxglove/studio";
@@ -18,6 +18,7 @@ import { PNG_TEST_IMAGE, rad2deg, SENSOR_FRAME_ID } from "../common";
 export default {
   title: "panels/ThreeDeeRender/Images",
   component: ThreeDeePanel,
+  parameters: { colorScheme: "light" },
 };
 
 const ImageModeRosImage = ({ imageType }: { imageType: "raw" | "png" }) => {
@@ -164,12 +165,10 @@ const ImageModeRosImage = ({ imageType }: { imageType: "raw" | "png" }) => {
 
 export const ImageModeRosRawImage: StoryObj = {
   render: () => <ImageModeRosImage imageType="raw" />,
-  parameters: { colorScheme: "light" },
 };
 
 export const ImageModeRosPngImage: StoryObj = {
   render: () => <ImageModeRosImage imageType="png" />,
-  parameters: { colorScheme: "light" },
 };
 
 const ImageModeFoxgloveImage = ({ imageType }: { imageType: "raw" | "png" }) => {
@@ -316,17 +315,14 @@ const ImageModeFoxgloveImage = ({ imageType }: { imageType: "raw" | "png" }) => 
 
 export const ImageModeFoxgloveRawImage: StoryObj = {
   render: () => <ImageModeFoxgloveImage imageType="raw" />,
-  parameters: { colorScheme: "light" },
 };
 
 export const ImageModeFoxglovePngImage: StoryObj = {
   render: () => <ImageModeFoxgloveImage imageType="png" />,
-  parameters: { colorScheme: "light" },
 };
 
 export const ImageModeResizeHandled: StoryObj = {
   render: () => <ImageModeFoxgloveImage imageType="raw" />,
-  parameters: { colorScheme: "light" },
 
   play: async () => {
     const canvas = document.querySelector("canvas")!;
@@ -339,9 +335,27 @@ export const ImageModeResizeHandled: StoryObj = {
   },
 };
 
+export const ImageModePanAndZoom: StoryObj = {
+  render: () => <ImageModeFoxgloveImage imageType="raw" />,
+  play: async () => {
+    const canvas = document.querySelector("canvas")!;
+    fireEvent.mouseDown(canvas, { clientX: 200, clientY: 200 });
+    fireEvent.mouseMove(canvas, { clientX: 400, clientY: 200 });
+    fireEvent.mouseUp(canvas, { clientX: 400, clientY: 200 });
+    fireEvent.wheel(canvas, { deltaY: 100, clientX: 200, clientY: 0 });
+  },
+};
+
+export const ImageModePanAndZoomReset: StoryObj = {
+  render: () => <ImageModeFoxgloveImage imageType="raw" />,
+  play: async (ctx) => {
+    await ImageModePanAndZoom.play?.(ctx);
+    userEvent.click(await screen.findByTestId("reset-view"));
+  },
+};
+
 export const ImageModePick: StoryObj = {
   render: () => <ImageModeFoxgloveImage imageType="raw" />,
-  parameters: { colorScheme: "light" },
 
   play: async () => {
     const canvas = document.querySelector("canvas")!;
